Only constrain to oldest numpy if not constraining scipy
There are only a limited number of scipy/numpy combinations
available for any given Python version. We could work out what
they all are, but it would be non-trivial to find the optimal
combination. Besides which, this does not seem so important
since only a narrow range of similarly aged numpy versions are
available for a given scipy version, so constraining scipy alone
will get a numpy version in the right ball-park.

Problems could arise if the user constrains the numpy version in
requirements.txt (e.g. numpy>=1.9.0), but does not constrain the
scipy version. This is pretty much unavoidable, and the solution
is for the user to constrain the scipy version so it is always
high enough for there to be a compatible numpy version which meets
the constraint.