CI: Change to bleeding-edge pre-commit/action
Updates https://github.com/pre-commit/action to use the master
version.

This closes a potential security vulnerability, which was addressed
in 80db042. We can not move to a later release, because we were
already on the newest release (v2.0.3). Moving to the bleeding-edge
(master branch) version of this action should not break anything
since the action is now in maintenance-only mode, and will not be
adding any new features.

Closes #95.